Reintroduce IT tests with OLLAMA
