+++ components/camel-ai/camel-langchain4j-tools/src/test/java/org/apache/camel/component/langchain4j/tools/LangChain4jToolMultipleGroupsTest.java
@@ -30,24 +30,29 @@ import org.apache.camel.test.infra.openai.mock.OpenAIMock;
 import org.apache.camel.test.junit5.CamelTestSupport;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.RepeatedTest;
-import org.junit.jupiter.api.condition.DisabledIfSystemProperty;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-@DisabledIfSystemProperty(named = "ci.env.name", matches = ".*", disabledReason = "Requires too much network resources")
-public class LangChain4jToolNoToolsExistIT extends CamelTestSupport {
+public class LangChain4jToolMultipleGroupsTest extends CamelTestSupport {
 
-    private ChatModel chatModel;
+    private final String nameFromDB = "pippo";
+    protected ChatModel chatModel;
 
     @RegisterExtension
     static OpenAIMock openAIMock = new OpenAIMock().builder()
-            .when("How can you help? DO NOT CALL TOOLS.\n")
+            .when("What is the name of the user 1?\n")
             .assertRequest(request -> {
-                // The tools should not be included in the request
-                Assertions.assertThat(request).doesNotContain(
+                // The tools has to be included in the request
+                Assertions.assertThat(request).contains(
                         "QueryUserDatabaseByNumber",
-                        "DoesNotDoAnything,Really");
+                        "DoesNotReallyDoAnything");
+
+                // The NOT included in the request
+                Assertions.assertThat(request).doesNotContain(
+                        "QueryCompanyUserDatabaseByNumber",
+                        "QuerySomethingelseUserDatabaseByNumber");
             })
-            .replyWith("No tool is invoked")
+            .invokeTool("{\"name\": \"pippo\"}")
+            .withParam("number", "1")
             .build();
 
     @Override
@@ -73,40 +78,41 @@ public class LangChain4jToolNoToolsExistIT extends CamelTestSupport {
     protected RouteBuilder createRouteBuilder() {
         return new RouteBuilder() {
             public void configure() {
+
                 from("direct:test")
-                        .to("langchain4j-tools:test1?tags=other")
+                        .to("langchain4j-tools:test1?tags=user")
                         .log("response is: ${body}");
 
                 from("langchain4j-tools:test1?tags=user&description=Query user database by number&parameter.number=integer")
                         .setBody(simple("{\"name\": \"pippo\"}"));
 
-                from("langchain4j-tools:test1?tags=user&description=Does not do anything, really")
-                        .setBody(simple("{\"name\": \"pippo\"}"));
+                from("langchain4j-tools:test1?tags=user&description=Does not really do anything")
+                        .setBody(constant("Hello World"));
 
-                from("direct:noResponse")
-                        .log("there is no tool to be called for the request: ${body}")
-                        .setBody(constant("There was no tool to be called"))
-                        .to("mock:noResponse");
+                from("langchain4j-tools:test1?tags=companies&description=Query company user database by number&parameter.number=integer")
+                        .setBody(constant("Hello World"));
 
+                from("langchain4j-tools:test1?tags=somethingelse&description=Query somethingelse user database by number&parameter.number=integer")
+                        .setBody(constant("Hello World"));
             }
         };
     }
 
-    @RepeatedTest(1)
-    public void testSimpleInvocation() throws InterruptedException {
+    @RepeatedTest(10)
+    public void testSimpleInvocation() {
         List<ChatMessage> messages = new ArrayList<>();
         messages.add(new SystemMessage(
                 """
-                        Your job is to help me test my code. When asked to call a tool you do not call anything.
+                        You provide the requested information using the functions you hava available. You can invoke the functions to obtain the information you need to complete the answer.
                         """));
         messages.add(new UserMessage("""
-                How can you help? DO NOT CALL TOOLS.
+                What is the name of the user 1?
                 """));
 
         Exchange message = fluentTemplate.to("direct:test").withBody(messages).request(Exchange.class);
 
         Assertions.assertThat(message).isNotNull();
-        Assertions.assertThat(message.getMessage().getHeader(LangChain4jTools.NO_TOOLS_CALLED_HEADER)).isEqualTo(Boolean.TRUE);
-
+        final String responseContent = message.getMessage().getBody().toString();
+        Assertions.assertThat(responseContent).containsIgnoringCase(nameFromDB);
     }
 }
