+++ components/camel-ai/camel-langchain4j-tools/src/test/java/org/apache/camel/component/langchain4j/tools/LangChain4jToolNoToolsExistTest.java
@@ -29,32 +29,23 @@ import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.test.infra.openai.mock.OpenAIMock;
 import org.apache.camel.test.junit5.CamelTestSupport;
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.condition.DisabledIfSystemProperty;
+import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-@DisabledIfSystemProperty(named = "ci.env.name", matches = ".*", disabledReason = "Requires too much network resources")
-public class LangChain4jToolMultipleMatchingGroupsIT extends CamelTestSupport {
+public class LangChain4jToolNoToolsExistTest extends CamelTestSupport {
 
-    private final String nameFromDB = "Paul McFly";
-    private ChatModel chatModel;
+    protected ChatModel chatModel;
 
     @RegisterExtension
     static OpenAIMock openAIMock = new OpenAIMock().builder()
-            .when("What is the name of the user 1 and what department he is part of?\n")
+            .when("How can you help? DO NOT CALL TOOLS.\n")
             .assertRequest(request -> {
-                // The tools has to be included in the request
-                Assertions.assertThat(request).contains(
-                        "QueryUserByNumberUsingTheInternalDB",
-                        "QueryUserInformationByNumberUsingTheExternalDB");
-
-                // The NOT included in the request
+                // The tools should not be included in the request
                 Assertions.assertThat(request).doesNotContain(
-                        "DoesNotReallyDoAnything",
-                        "QueryCarInformationByPlate");
+                        "QueryUserDatabaseByNumber",
+                        "DoesNotDoAnything,Really");
             })
-            .invokeTool("{\"name\": \"Paul McFly\", \"department\": \"engineering\"}")
-            .withParam("number", 1)
+            .replyWith("No tool is invoked")
             .build();
 
     @Override
@@ -80,42 +71,40 @@ public class LangChain4jToolMultipleMatchingGroupsIT extends CamelTestSupport {
     protected RouteBuilder createRouteBuilder() {
         return new RouteBuilder() {
             public void configure() {
-
                 from("direct:test")
-                        .to("langchain4j-tools:test1?tags=internalDB,externalDB")
+                        .to("langchain4j-tools:test1?tags=other")
                         .log("response is: ${body}");
 
-                from("langchain4j-tools:test2?tags=internalDB&description=Query user by number using the internal DB&parameter.number=integer")
-                        .setBody(simple("{\"name\": \"Paul McFly\"}"));
+                from("langchain4j-tools:test1?tags=user&description=Query user database by number&parameter.number=integer")
+                        .setBody(simple("{\"name\": \"pippo\"}"));
 
-                from("langchain4j-tools:test3?tags=notApplicable&description=Does not really do anything")
-                        .setBody(constant("Hello World"));
+                from("langchain4j-tools:test1?tags=user&description=Does not do anything, really")
+                        .setBody(simple("{\"name\": \"pippo\"}"));
 
-                from("langchain4j-tools:test4?tags=externalDB&description=Query user information by number using the external DB&parameter.number=integer")
-                        .setBody(simple("{\"name\": \"Paul McFly\", \"department\": \"engineering\"}"));
+                from("direct:noResponse")
+                        .log("there is no tool to be called for the request: ${body}")
+                        .setBody(constant("There was no tool to be called"))
+                        .to("mock:noResponse");
 
-                from("langchain4j-tools:test5?tags=notApplicable&description=Query car information by plate&parameter.number=integer")
-                        .setBody(constant("Hello World"));
             }
         };
     }
 
-    @Test
-    public void testSimpleInvocation() {
+    @RepeatedTest(1)
+    public void testSimpleInvocation() throws InterruptedException {
         List<ChatMessage> messages = new ArrayList<>();
         messages.add(new SystemMessage(
                 """
-                        You provide the requested information using the functions you hava available. You can invoke the functions to obtain the information you need to complete the answer.
+                        Your job is to help me test my code. When asked to call a tool you do not call anything.
                         """));
         messages.add(new UserMessage("""
-                What is the name of the user 1 and what department he is part of?
+                How can you help? DO NOT CALL TOOLS.
                 """));
 
         Exchange message = fluentTemplate.to("direct:test").withBody(messages).request(Exchange.class);
 
         Assertions.assertThat(message).isNotNull();
-        final String responseContent = message.getMessage().getBody().toString();
-        Assertions.assertThat(responseContent).containsIgnoringCase(nameFromDB);
-        Assertions.assertThat(responseContent).containsIgnoringCase("engineering");
+        Assertions.assertThat(message.getMessage().getHeader(LangChain4jTools.NO_TOOLS_CALLED_HEADER)).isEqualTo(Boolean.TRUE);
+
     }
 }
