+++ components/camel-ai/camel-langchain4j-tools/src/test/java/org/apache/camel/component/langchain4j/tools/ToolsHelper.java
@@ -22,7 +22,7 @@ import dev.langchain4j.model.openai.OpenAiChatModel;
 
 import static java.time.Duration.ofSeconds;
 
-final class ToolsHelper {
+public final class ToolsHelper {
 
     private ToolsHelper() {
     }
